refactor!: update the type of total count from uint32 to int64
BREAKING CHANGE: The TotalCount field of BaseWithTotalCountResponse DTO will be changed from uint32 to int64

Since postgresql is the default persistence layer and the count() func returned from postgresql will be
bigint(https://www.postgresql.org/docs/17/functions-aggregate.html) with value range int64, we should
update the total count value type from uint32 to int64 to avoid integer overflow issue, where the max
of uint32 is 4,294,967,295 and the max of int64 is 9,223,372,036,854,775,807.

+++ dtos/common/base.go
@@ -66,10 +66,10 @@ func NewBaseWithIdResponse(requestId string, message string, statusCode int, id 
 // BaseWithTotalCountResponse defines the base content for response DTOs (data transfer objects).
 type BaseWithTotalCountResponse struct {
 	BaseResponse `json:",inline"`
-	TotalCount   uint32 `json:"totalCount"`
+	TotalCount   int64 `json:"totalCount"`
 }
 
-func NewBaseWithTotalCountResponse(requestId string, message string, statusCode int, totalCount uint32) BaseWithTotalCountResponse {
+func NewBaseWithTotalCountResponse(requestId string, message string, statusCode int, totalCount int64) BaseWithTotalCountResponse {
 	return BaseWithTotalCountResponse{
 		BaseResponse: NewBaseResponse(requestId, message, statusCode),
 		TotalCount:   totalCount,
