+++ dtos/responses/reading_test.go
@@ -35,7 +35,7 @@ func TestNewMultiReadingsResponse(t *testing.T) {
 		{Id: "7a1707f0-166f-4c4b-bc9d-1d54c74e0137"},
 		{Id: "11111111-2222-3333-4444-555555555555"},
 	}
-	expectedTotalCount := uint32(len(expectedReadings))
+	expectedTotalCount := int64(len(expectedReadings))
 	actual := NewMultiReadingsResponse(expectedRequestId, expectedMessage, expectedStatusCode, expectedTotalCount, expectedReadings)
 
 	assert.Equal(t, expectedRequestId, actual.RequestId)
